# api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: HireMind AI API Spec
  description: Spécification des contrats d'API pour la plateforme SaaS de recrutement HireMind AI. Ce document définit l'interface d'intégration entre le Client Mobile (Flutter), le Portail Web (Next.js 15) et le Core Backend (NestJS).
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Serveur de développement local NestJS
paths:
...
  /auth/register:
    post:
      summary: Inscription d'un utilisateur
      description: Permet d'enregistrer un candidat ou un recruteur.
      tags:
        - Authentification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - role
              properties:
                email:
                  type: string
                  format: email
                  example: candidate@hiremind.ai
                password:
                  type: string
                  format: password
                  example: "SecurePass123!"
                role:
                  type: string
                  enum: [candidate, recruiter, admin]
                  example: candidate
      responses:
        '201':
          description: Utilisateur créé avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "usr_98237498"
                  email:
                    type: string
                    example: candidate@hiremind.ai
                  role:
                    type: string
                    example: candidate
        '400':
          description: Données invalides ou email déjà utilisé.
  /auth/login:
    post:
      summary: Connexion utilisateur
      description: Authentifie l'utilisateur et retourne un token JWT.
      tags:
        - Authentification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: candidate@hiremind.ai
                password:
                  type: string
                  format: password
                  example: "SecurePass123!"
      responses:
        '200':
          description: Authentification réussie.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  expiresIn:
                    type: integer
                    example: 3600
        '401':
          description: Identifiants invalides.
  /auth/me:
    get:
      summary: Obtenir le profil de l'utilisateur connecté
      description: Retourne les détails du profil de l'utilisateur à partir du token JWT.
      tags:
        - Authentification
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profil utilisateur retourné.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "usr_98237498"
                  email:
                    type: string
                    example: candidate@hiremind.ai
                  role:
                    type: string
                    example: candidate
                  profile:
                    type: object
                    properties:
                      firstName:
                        type: string
                        example: "Jean"
                      lastName:
                        type: string
                        example: "Dupont"
        '401':
          description: Non authentifié.

  # --- CV & PASS SPORT IA ---
  /cv/upload:
    post:
      summary: Déposer un CV pour parsing automatique
      description: Importe un fichier de CV (PDF, DOCX, TXT), extrait les informations avec un LLM et retourne les données structurées.
      tags:
        - CV & Parsing
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Fichier du CV (Max 5MB).
      responses:
        '201':
          description: CV importé et analysé avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  cvUrl:
                    type: string
                    format: uri
                    example: "https://minio.hiremind.internal/cvs/raw_cv_98237498.pdf"
                  parsedData:
                    type: object
                    properties:
                      identity:
                        type: object
                        properties:
                          fullName:
                            type: string
                            example: "Jean Dupont"
                          email:
                            type: string
                            example: "jean.dupont@email.com"
                          phone:
                            type: string
                            example: "+33612345678"
                      skills:
                        type: object
                        properties:
                          technical:
                            type: array
                            items:
                              type: string
                            example: ["Flutter", "Dart", "Firebase", "Git"]
                          methodological:
                            type: array
                            items:
                              type: string
                            example: ["Agile", "Scrum", "TDD", "Clean Architecture"]
                          softSkills:
                            type: array
                            items:
                              type: string
                            example: ["Communication", "Leadership", "Esprit d'équipe"]
                      experiences:
                        type: array
                        items:
                          type: object
                          properties:
                            company:
                              type: string
                              example: "Tech Corp"
                            role:
                              type: string
                              example: "Développeur Mobile"
                            duration:
                              type: string
                              example: "2 ans"
                            description:
                              type: string
                              example: "Développement d'applications Flutter pour le secteur bancaire."
  /candidates/me/passport:
    get:
      summary: Obtenir le Passeport de Compétences (Skill Passport IA)
      description: Récupère les données structurées du passeport de compétences pour l'affichage en graphique radar.
      tags:
        - CV & Parsing
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Passeport de compétences récupéré.
          content:
            application/json:
              schema:
                type: object
                properties:
                  candidateId:
                    type: string
                    example: "usr_98237498"
                  radarScores:
                    type: array
                    items:
                      type: object
                      properties:
                        axis:
                          type: string
                          description: Axe du graphique radar (ex. Technique, Soft Skills, Gestion de projet)
                          example: "Technique"
                        score:
                          type: integer
                          description: Score sur 100
                          example: 85

  # --- OFFRES D'EMPLOI & MATCHING ---
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT